feat(formula-bar): track recent name box selections
- Add in-memory 'Recent' section to the name box listbox dropdown\n  (most recently selected targets bubble to the top).\n- Add jsdom coverage for recents behavior.\n- Fix ai-audit stable-json bigint stringify TS narrowing.

+++ packages/ai-audit/src/stable-json.ts
@@ -34,7 +34,8 @@ export function stableJsonValue(value: unknown, ancestors: WeakSet<object>): unk
 
   const t = typeof value;
   if (t === "string" || t === "number" || t === "boolean") return value;
-  if (t === "bigint") return value.toString();
+  // `t` is derived from `typeof value`, but TypeScript can't use that to narrow `value` here.
+  if (t === "bigint") return (value as bigint).toString();
   if (t === "undefined" || t === "function" || t === "symbol") return undefined;
 
   if (Array.isArray(value)) {
@@ -94,4 +95,3 @@ export function stableJsonValue(value: unknown, ancestors: WeakSet<object>): unk
   ancestors.delete(obj);
   return sorted;
 }
-
